Fix critical bugs: uninitialized vars, use-after-free, type errors, domain errors, typos, thread safety
- target.cpp: Initialize `id` to 0 to prevent undefined behavior
- aimer.hpp: Change `lock_id_` from double to int (correct type for array index)
- buff_aimer.cpp: Fix `fsqrt()` → `std::sqrt()`, fix uninitialized yaw/pitch usage
- mt_detector.cpp: Fix use-after-free by using ov::Tensor with owned memory
- math_tools.cpp: Clamp acos() arguments to [-1,1] to prevent NaN
- admm.cpp: Fix `abs()` → `std::abs()` for float precision
- detector.cpp: Guard cv::imshow with debug_ flag for headless environments
- thread_pool.hpp: Add mutex lock to get_size(), fix typo numebr→number
- thread_safe_queue.hpp: Remove unreachable dead code after condition_variable wait
- solver.hpp/cpp: Fix typos oupost→outpost, picth→pitch

+++ tasks/auto_aim/aimer.hpp
@@ -37,7 +37,7 @@ private:
   double pitch_offset_;
   double comming_angle_;
   double leaving_angle_;
-  double lock_id_ = -1;
+  int lock_id_ = -1;
   double high_speed_delay_time_;
   double low_speed_delay_time_;
   double decision_speed_;
